Add event to 'child' parsers
It seems to work like this: there's an event when you see a child test
which gives you a parser for that section of TAP. You need to add events
to that newly created parser to see what's happening in child tests.

Also .gitignore the generated content-script-bundle.js from browserify.