Use a promise for async UI
Updating a long TAP document is slow, so we can put that task in a
setTimeout so that the UI gets a chance to draw the toolbar before
moving on to the UI. This gives the user a clue that something is
working.

Unfortunately this breaks the tests, as control returns to them before
the work has been done. We can get round this by having processDocument
return a Promise, which can then be chained with then() in the test so
that the expectation is not examined until there is a result.